Refactor MetricsController
