+++ core/src/main/java/eu/tsystems/mms/tic/testframework/internal/metrics/MetricsType.java
@@ -1,0 +1,13 @@
+package eu.tsystems.mms.tic.testframework.internal.metrics;
+
+/**
+ * Created on 2023-10-16
+ *
+ * @author mgn
+ */
+public enum MetricsType {
+    BASEURL_LOAD,
+    SESSION_LOAD,
+    SESSION_DURATION,
+    METHOD_DURATION;
+}
